wip 一旦動く状態に再構築
