+++ frontend/fresh.gen.ts
@@ -4,27 +4,17 @@
 
 import * as $0 from "./routes/[name].tsx";
 import * as $1 from "./routes/api/joke.ts";
-import * as $2 from "./routes/articles/[id].tsx";
-import * as $3 from "./routes/articles/create.tsx";
-import * as $4 from "./routes/index.tsx";
-import * as $5 from "./routes/search.tsx";
-import * as $6 from "./routes/test.tsx";
-import * as $$0 from "./islands/ContentForm.tsx";
-import * as $$1 from "./islands/Counter.tsx";
+import * as $2 from "./routes/index.tsx";
+import * as $$0 from "./islands/Counter.tsx";
 
 const manifest = {
   routes: {
     "./routes/[name].tsx": $0,
     "./routes/api/joke.ts": $1,
-    "./routes/articles/[id].tsx": $2,
-    "./routes/articles/create.tsx": $3,
-    "./routes/index.tsx": $4,
-    "./routes/search.tsx": $5,
-    "./routes/test.tsx": $6,
+    "./routes/index.tsx": $2,
   },
   islands: {
-    "./islands/ContentForm.tsx": $$0,
-    "./islands/Counter.tsx": $$1,
+    "./islands/Counter.tsx": $$0,
   },
   baseUrl: import.meta.url,
 };
