+++ frontend/import_map.json
@@ -1,18 +1,11 @@
 {
   "imports": {
-    "$fresh/": "https://raw.githubusercontent.com/lucacasonato/fresh/main/",
-    "preact": "https://esm.sh/preact@10.8.1",
-    "preact/": "https://esm.sh/preact@10.8.1/",
-    "preact-render-to-string": "https://esm.sh/preact-render-to-string@5.2.0?deps=preact@10.8.1",
+    "$fresh/": "https://deno.land/x/fresh@1.0.1/",
+    "preact": "https://esm.sh/preact@10.8.2",
+    "preact/": "https://esm.sh/preact@10.8.2/",
+    "preact-render-to-string": "https://esm.sh/preact-render-to-string@5.2.0?deps=preact@10.8.2",
     "@twind": "./utils/twind.ts",
     "twind": "https://esm.sh/twind@0.16.17",
-    "twind/": "https://esm.sh/twind@0.16.17/",
-    "dayjs/": "https://esm.sh/dayjs@1.11.3/",
-    "@db": "./utils/db.ts",
-    "postgress": "https://deno.land/x/postgres@v0.16.1/mod.ts",
-    "dotenv/": "https://deno.land/x/dotenv@v3.2.0/",
-    "marked": "https://esm.sh/marked@4.0.17",
-    "sanitize-html": "https://esm.sh/sanitize-html@2.7.0",
-    "dompurify": "https://esm.sh/dompurify@2.3.8"
+    "twind/": "https://esm.sh/twind@0.16.17/"
   }
 }
