+++ frontend/routes/index.tsx
@@ -1,99 +1,41 @@
 /** @jsx h */
-import { h, PageProps } from "preact";
+import { h } from "preact";
 import { Head } from "$fresh/src/runtime/head.ts";
-import { Handlers } from "$fresh/server.ts";
 import { tw } from "@twind";
 
-// dayjs で相対日付を表示するために import
-import dayjs from "dayjs/";
-import relativeTime from "dayjs/plugin/relativeTime";
-import "dayjs/locale/ja";
-// カスタムハンドラー
-import { Article, findAllArticles } from "@db";
-
-dayjs.extend(relativeTime);
-dayjs.locale("ja");
-
-interface Article {
-  id: string;
-  title: string;
-  created_at: string;
-}
-
-export const handler: Handlers<Article[]> = {
-  async GET(_, ctx) {
-    const articles = await findAllArticles();
-    return ctx.render(articles);
-  },
-};
-// export const handler: Handlers<Article[]> = {
-//   async GET(_, ctx) {
-//     const articles: Article[] = [
-//       {
-//         id: "1",
-//         title: "しぐれうい先生のメン限が良かった",
-//         created_at: "2022-06-29T00:00:00.000Z",
-//       },
-//       {
-//         id: "2",
-//         title: "GCP何もわからない",
-//         created_at: "2022-06-30T00:00:00.000Z",
-//       },
-//     ];
-//     return ctx.render(articles);
-//   },
-// };
-export default function Home({ data }: PageProps<Article[]>) {
+export default function Home() {
   return (
-    <div class={tw("h-screen bg-gray-200")}>
+    <div class={tw("h-screen")}>
       <Head>
         <title>Usuyuki Blog</title>
+        <link rel="stylesheet" href="/n.css" />
       </Head>
       <div
         class={tw(
           "max-w-screen-sm mx-auto px-4 sm:px-6 md:px-8 pt-12 pb-20 flex flex-col"
         )}
       >
-        <h1 class={tw("font-extrabold text-5xl text-gray-800")}>
-          Usuyuki Blog
-        </h1>
-        <section class={tw("mt-8")}>
-          <div class={tw("flex justify-between items-center")}>
-            <h2 class={tw("text-4xl font-bold text-gray-800 py-4")}>Posts</h2>
-            <a
-              href="/articles/create"
-              class={tw(
-                "bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-md"
-              )}
-            >
-              Create Post
-            </a>
+        <div class={tw("absolute")}>
+          <div class={tw("relative top-4 left-20")}>
+            <div class="glass-frame">
+              <p class={"kaisei-decol " + tw("text-2xl text-center mt-20")}>
+                サイン
+              </p>
+            </div>
           </div>
-          <ul>
-            {data.map((article) => (
-              <li
-                class={tw("bg-white p-6 rounded-lg shadow-lg mb-4")}
-                key={article.id}
-              >
-                <a href={`articles/${article.id}`}>
-                  <h3
-                    class={tw(
-                      "text-2xl font-bold mb-2 text-gray-800 hover:text-gray-600 hover:text-underline"
-                    )}
-                  >
-                    {article.title}
-                  </h3>
-                </a>
-                <time
-                  class={tw("text-gray-500 text-sm")}
-                  dateTime={article.created_at}
-                >
-                  {dayjs(article.created_at).fromNow()}
-                </time>
-              </li>
-            ))}
-          </ul>
-        </section>
+          <div class="glass-frame">
+            <p class={"kaisei-decol " + tw("text-2xl text-center mt-20")}>
+              コサイン
+            </p>
+          </div>
+          <div class={tw("relative bottom-4 right-20")}>
+            <div class="glass-frame">
+              <p class={"kaisei-decol " + tw("text-2xl text-center mt-20")}>
+                ですわ～
+              </p>
+            </div>
+          </div>
+        </div>
       </div>
     </div>
   );
